requirements.txt
numpy==1.26.4
timm==0.9.16
diffusers==0.37.0
transformers==4.53.3
peft==0.17.1
safetensors==0.5.3
datasets==2.21.0
accelerate==1.13.0
huggingface_hub==0.34.4
sentencepiece
insightface
onnx==1.19.1
#--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-11/pypi/simple/
#onnxruntime-gpu==1.20.1
onnxruntime-gpu==1.23.2
git+https://github.com/FacePerceiver/facer.git@main
git+https://github.com/openai/CLIP.git
#git+https://github.com/facebookresearch/pytorch3d.git@v0.7.9
#chumpy
# pip install chumpy==0.70 --no-build-isolation
einops
protobuf
opencv-python==4.11.0.86
tensorboard
torch_tps
fvcore
albumentations
imgaug
timed-decorator
torch-fidelity
torchmetrics
openai
google-genai
rotary-embedding-torch
ninja
#tensorflow==2.10.1
# pip install xformers --index-url https://download.pytorch.org/whl/cu118
# git clone --recurse-submodules https://github.com/facebookresearch/xformers
# pip install -v --no-build-isolation .
